SMTChecker: Rename CVC interface files
